requirements.txt
requests
bs4==0.0.1
beautifulsoup4
pyquery==1.4.0
ruamel.yaml==0.16.13
GitPython==3.1.24
Telethon